job_application_agent/requirements.txt
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.4
schedule==1.2.0
python-dotenv==1.0.0
webdriver-manager==4.0.1
fake-useragent==1.4.0
lxml==4.9.3
openpyxl==3.1.2
sqlalchemy==2.0.23